requirements.txt
Django==4.1
django-filter==23.4
mysqlclient==2.2.0
Pillow==10.1.0
python-dotenv==1.0.0
sorl-thumbnail==12.10.0
django-admin-sortable2==2.1.10

